Fix memory reclaimation bug in Resource Manager Service.
When insufficient resource error is reported at configure
stage, current OMX instance is registered as a resource
manager client. When determing which codec instance to release,
there is a comparison between the priority of all registered
clients and current running instance. When current instance is
a secure codec and multiple secure decoders are not supported, the
policy is to kick out other secure instance as the first step to
reclaim the resource. The problem is when comparing the priority,
it always compares to itself first and since current client is also
registered the comparision always fails. As a result, the
function simply returns and resource reclaimation stops. The fix
is to allow the reclaiming procedure continue to work on other
non-secure codecs.

Test: Build is successful, atest CtsVideoTestCases, atest
mediacodecTest, atest CtsMediaMiscTestCases, atest
CtsMediaPlayerTestCases

Change-Id: If430d8184bd5efa274e0cff62a97697801dd7c1c
Signed-off-by: Saurabh <sbbanore@amazon.com>