Stop serving HIDL media.c2 software instance in 202404 devices
New launching devices targeting 202404 vendor interface will not use
HIDL, so mediaserver can stop serving it.
max-level="8" means that the entry is removed from the manifest on
devices targeting vendor interfaces after 8 (Android U), so
hwservicemanager will now prevent it from being registered on those
devices.
Mediaserver checks to see if the HIDL service is declared, if it is
declared then it throws an error when there is an issue with registering
the service with hwservicemanager. If it is not declared, it does not
try to register the service.

Test: atest CtsMediaCodecTestCases
Bug: 218588089
Change-Id: I97858271ef60e4c11e4fce28acc68439b31a49ec